Fix for #26 : Reuse existing bundles option while importing templates
There is now a new option --rbundles on the following commands :
- hammr template create
- hammr template import

When this option is set, if you have a bundle described in your template file and if the bundle exists on the ser
ver, the server version will be attached to the newly created/imported template.

Warning : the server version of the bundle might not contain exactly the same content (files, content of the file
s, etc.). The UForge server uses the key : user, bundle name, bundle version to know if the bundle already exists.